Fix TE_OFFSET_OF
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@29382 9d14d920-75d5-0310-8bd7-a52572981aee

+++ include/te_defs.h
@@ -179,7 +179,7 @@ typedef unsigned char te_bool;
  * @param _s        Structure name
  * @param _f        Field name
  */
-#define TE_OFFSET_OF(_s, _f)    (&(((_s *)NULL)->_f))
+#define TE_OFFSET_OF(_s, _f)    ((size_t)(&(((_s *)NULL)->_f)))
 
 /** 
  * Number of elements in array.
